requirements.txt
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
cryptography==35.0.0
distro==1.6.0
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
google==3.0.0
idna==2.10
jsonschema==3.2.0
paramiko==2.8.0
pycparser==2.21
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.3
texttable==1.6.4
urllib3==1.26.7
websocket-client==0.59.0
